LibTTF: Reorganized TTF parsing code into LibTTF
