+++ Libraries/CMakeLists.txt
@@ -10,12 +10,12 @@ add_subdirectory(LibCrypto)
 add_subdirectory(LibDebug)
 add_subdirectory(LibDesktop)
 add_subdirectory(LibDiff)
-add_subdirectory(LibGUI)
 add_subdirectory(LibGemini)
 add_subdirectory(LibGfx)
+add_subdirectory(LibGUI)
 add_subdirectory(LibHTTP)
-add_subdirectory(LibIPC)
 add_subdirectory(LibImageDecoderClient)
+add_subdirectory(LibIPC)
 add_subdirectory(LibJS)
 add_subdirectory(LibKeyboard)
 add_subdirectory(LibLine)
@@ -25,10 +25,11 @@ add_subdirectory(LibPCIDB)
 add_subdirectory(LibProtocol)
 add_subdirectory(LibPthread)
 add_subdirectory(LibRegex)
-add_subdirectory(LibTLS)
 add_subdirectory(LibTar)
 add_subdirectory(LibTextCodec)
 add_subdirectory(LibThread)
+add_subdirectory(LibTLS)
+add_subdirectory(LibTTF)
 add_subdirectory(LibVT)
 add_subdirectory(LibWeb)
 add_subdirectory(LibX86)
