+++ Libraries/LibGfx/CMakeLists.txt
@@ -30,8 +30,7 @@ set(SOURCES
     SystemTheme.cpp
     Triangle.cpp
     WindowTheme.cpp
-    TTFont.cpp
 )
 
 serenity_lib(LibGfx gfx)
-target_link_libraries(LibGfx LibM LibCore LibIPC)
+target_link_libraries(LibGfx LibM LibCore LibTTF)
