+++ Libraries/LibTTF/Font.h
@@ -28,12 +28,11 @@
 
 #include <AK/ByteBuffer.h>
 #include <AK/Noncopyable.h>
+#include <AK/Optional.h>
 #include <AK/RefCounted.h>
 #include <AK/StringView.h>
 #include <LibGfx/Bitmap.h>
 #include <LibGfx/Size.h>
-
-namespace Gfx {
 
 #define POINTS_PER_INCH 72.0f
 #define DEFAULT_DPI     96
@@ -82,7 +81,7 @@ private:
     Font(ByteBuffer&& buffer, u32 offset);
     ScaledFontMetrics metrics(float x_scale, float y_scale) const;
     ScaledGlyphMetrics glyph_metrics(u32 glyph_id, float x_scale, float y_scale) const;
-    RefPtr<Bitmap> raster_glyph(u32 glyph_id, float x_scale, float y_scale) const;
+    RefPtr<Gfx::Bitmap> raster_glyph(u32 glyph_id, float x_scale, float y_scale) const;
     u32 glyph_count() const { return m_maxp.num_glyphs(); }
 
     enum class IndexToLocFormat {
@@ -330,7 +329,7 @@ private:
         public:
             static Glyph simple(const ByteBuffer& slice, u16 num_contours, i16 xmin, i16 ymin, i16 xmax, i16 ymax);
             static Glyph composite(const ByteBuffer& slice); // FIXME: This is currently just a dummy. Need to add support for composite glyphs.
-            RefPtr<Bitmap> raster(float x_scale, float y_scale) const;
+            RefPtr<Gfx::Bitmap> raster(float x_scale, float y_scale) const;
             int ascender() const
             {
                 if (m_type == Type::Simple) {
@@ -368,7 +367,7 @@ private:
                 , m_slice(move(slice))
             {
             }
-            RefPtr<Bitmap> raster_simple(float x_scale, float y_scale) const;
+            RefPtr<Gfx::Bitmap> raster_simple(float x_scale, float y_scale) const;
 
             Type m_type;
             ByteBuffer m_slice;
@@ -423,7 +422,7 @@ public:
     u32 glyph_id_for_codepoint(u32 codepoint) const { return m_font->m_cmap.glyph_id_for_codepoint(codepoint); }
     ScaledFontMetrics metrics() const { return m_font->metrics(m_x_scale, m_y_scale); }
     ScaledGlyphMetrics glyph_metrics(u32 glyph_id) const { return m_font->glyph_metrics(glyph_id, m_x_scale, m_y_scale); }
-    RefPtr<Bitmap> raster_glyph(u32 glyph_id) const { return m_font->raster_glyph(glyph_id, m_x_scale, m_y_scale); }
+    RefPtr<Gfx::Bitmap> raster_glyph(u32 glyph_id) const { return m_font->raster_glyph(glyph_id, m_x_scale, m_y_scale); }
     u32 glyph_count() const { return m_font->glyph_count(); }
     int width(const StringView&) const;
     int width(const Utf8View&) const;
@@ -437,5 +436,4 @@ private:
     friend Font;
 };
 
-}
 }
